requirements.txt
pure-python-adb
opencv-python
pillow
numpy
openai
pytesseract
matplotlib
logging
setuptools
requests
uiautomator2
python-dotenv